+++ app/src/main/java/co/ghola/smogalert/chart/DayAxisValueFormatter.java
@@ -1,10 +1,14 @@
 package co.ghola.smogalert.chart;
 
+import android.util.Log;
+
 import com.github.mikephil.charting.charts.BarLineChartBase;
 import com.github.mikephil.charting.components.AxisBase;
 import com.github.mikephil.charting.formatter.AxisValueFormatter;
+import com.google.api.client.util.DateTime;
 
 import java.text.DecimalFormat;
+import java.util.ArrayList;
 
 /**
  * Created by alecksjohansson on 7/28/16.
@@ -19,13 +23,22 @@ public class DayAxisValueFormatter implements AxisValueFormatter {
 
     public DayAxisValueFormatter()
     {
-        mFormat = new DecimalFormat("###,###,##0.0");
+        mFormat = new DecimalFormat("###,###,##0");
     }
 
     @Override
     public String getFormattedValue(float value, AxisBase axis) {
-        int days = (int) value;
-        return days + " Day"; // e.g. append a dollar-sign
+        if((int) value == 7){
+
+            return "";
+        }
+        int days = (int) value * 4;
+        if(days<12){
+            return days+" AM";
+        }else {
+            return (days-12)+" PM";
+        }
+
     }
