+++ app/src/main/res/layout/content_main.xml
@@ -30,7 +30,7 @@
             <android.support.v4.view.ViewPager
                 android:id="@+id/vpPager"
                 android:layout_width="match_parent"
-                android:layout_height="151dp"
+                android:layout_height="147dp"
                 android:layout_gravity="top"
                 />
 
@@ -85,7 +85,7 @@
                 <android.support.v4.view.ViewPager
                     android:id="@+id/vpPager2"
                     android:layout_width="match_parent"
-                    android:layout_height="189dp"
+                    android:layout_height="180dp"
                     android:layout_gravity="top" />
