+++ app/src/main/res/layout/fouth_fragment.xml
@@ -19,7 +19,8 @@
         android:layout_height="wrap_content"
         android:id="@+id/location"
         android:textStyle="bold"
-        android:textColor="#FFF"/>
+        android:textColor="#FFF"
+        android:layout_marginTop="4dp" />
 
 
     <!-- Here the current weather data -->
@@ -30,7 +31,8 @@
         android:layout_height="@dimen/weatherIcon.height"
         android:id="@+id/tempIcon"
         android:src="@drawable/temperature"
-        android:layout_below="@id/location" />
+        android:layout_below="@id/location"
+        android:layout_marginTop="10dp" />
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -112,7 +114,8 @@
         android:src="@drawable/sun"
         android:layout_toRightOf="@id/windDeg"
         android:layout_alignBaseline="@id/tempIcon"
-        android:layout_below="@id/location" />
+        android:layout_below="@id/location"
+        android:layout_marginTop="10dp" />
 
     <TextView
         android:layout_width="wrap_content"
