+++ app/src/main/res/layout/third_fragment.xml
@@ -7,66 +7,76 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
     android:background="#3F51B5">
-
-
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:id="@+id/location"
         android:textStyle="bold"
         android:textColor="#FFF"/>
-
-    <RelativeLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/tempLyt"
-        android:layout_below="@id/location"
-        android:layout_centerHorizontal="true">
-
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             style="@style/textBig"
             android:id="@+id/temp"
-            />
+            android:layout_marginTop="15dp"
+            android:layout_marginLeft="23dp" />
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="match_parent"
+            android:id="@+id/tempUnit"
+            android:layout_alignBaseline="@id/temp"
+            android:layout_toRightOf="@id/temp"
+            style="@style/textVerySmall"
+            android:layout_marginTop="20dp" />
+    </LinearLayout>
+    <RelativeLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:id="@+id/tempLyt"
+        android:layout_centerHorizontal="true"
+        android:layout_alignParentRight="true"
+        android:layout_marginRight="20dp"
+        android:layout_below="@+id/location"
+        android:layout_marginTop="2dp">
+
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="3dp"
-            android:layout_alignLeft="@id/temp"
-            android:layout_alignRight="@id/temp"
+
             android:id="@+id/lineTxt"
             android:layout_below="@id/temp"
-
             android:layout_marginTop="0dp" />
 
         <ImageView
             android:layout_width="@dimen/weatherMainIcon.width"
             android:layout_height="@dimen/weatherMainIcon.height"
-            android:layout_marginLeft="10dp"
+            android:layout_marginLeft="15dp"
             android:id="@+id/imgWeather"
             android:layout_toRightOf="@id/temp"
             android:layout_alignTop="@id/temp"
             android:background="@drawable/circle"
-            android:padding="4dp"/>
+            android:padding="4dp"
+            android:layout_alignParentRight="true"
+            android:layout_marginRight="23dp"
+            android:layout_marginTop="3dp" />
 
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:id="@+id/tempUnit"
-            android:layout_alignBaseline="@id/temp"
-            android:layout_toRightOf="@id/temp"
-            android:layout_alignStart="@id/imgWeather"
-            style="@style/textVerySmall"/>
+
 
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:id="@+id/descrWeather"
             android:layout_below="@id/imgWeather"
-            android:layout_toRightOf="@id/temp"
             android:layout_alignStart="@id/tempUnit"
+            android:gravity="center_horizontal"
+
             style="@style/textSmall"
-            android:layout_marginTop="10dp" />
+            android:layout_marginTop="6dp"
+            android:layout_alignParentRight="true"
+            android:layout_centerHorizontal="true" />
 
 
     </RelativeLayout>
